LokalMarket reponsivo/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-mail==0.9.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
cryptography==41.0.7